Updated downloader
Removed the file clearing function at the end of the script and replaced it with a dynamic removal of links already downloaded. As a file is downloaded, it is removed from the "Links.txt" file.

Also included a 'Limit' to the amount of downloads. In this case 1000 links at a time. This allows the script or complied app, to be scheduled and will only fetch a certain amount, 1000, of records each run.

Currently using this on a Windows machine with a scheduled task that runs a compiled exe of this script every hour after startup. This allows control of when and how much downloading is done at a time. Bulk links can then be downloaded during normal operation and does not require the system to remain on to process all of the requests, it will get there, eventually.

Lastly, for any links that produced an error, an error file will be created with the links that caused the download error, if any.