First upload of the Split list file.
Initial upload.

This program was created to quickly split URL links lists. The URL links are placed in a file called "Links.txt" and the program should then be executed. Once executed the program will try and create 8 equal length list files (1 - 8.txt) with an equal amount of records for download. The last file might be a record or so short based on whether the Links.txt file record count can be evenly split between 8 files or not. The application reads the file, and pops off the last record into the new file(s) until all records have been split between them. This can then be used with the bulk image downloader in multiple instances to actually do the downloads.