Add files via upload
Added the PyTorch cuda capable image calssifier self trained and maintained.

The pretrained model was trained on 3 classes. ['nude', 'safe','sexy']

The idea is to create a final classifier that can distinguish between SFW (Suitable for work) and NSFW (Not suitable for work) content and then move the files into respective subfolders.